refactor: extract community posts hook
